+++ EasyVCR/AdvancedSettings.cs
@@ -1,6 +1,9 @@
 using System;
 using EasyVCR.Interfaces;
 using EasyVCR.InternalUtilities;
+using Microsoft.Extensions.Logging;
+
+// ReSharper disable FieldCanBeMadeReadOnly.Global
 
 namespace EasyVCR
 {
@@ -22,6 +25,11 @@ namespace EasyVCR
         public IInteractionConverter? InteractionConverter = new DefaultInteractionConverter();
 
         /// <summary>
+        ///     Tell EasyVCR to use a custom <see cref="ILogger"/> logger. Will log to console otherwise.
+        /// </summary>
+        public ILogger? Logger = null;
+
+        /// <summary>
         ///     Simulate a delay in milliseconds for the request.
         /// </summary>
         public int ManualDelay = 0;
@@ -34,6 +42,16 @@ namespace EasyVCR
         ///     Simulate the original request's duration. Overrides <see cref="ManualDelay"/> if set.
         /// </summary>
         public bool SimulateDelay = false;
+
+        /// <summary>
+        ///     How long should a given cassette entry be considered valid.
+        /// </summary>
+        public TimeFrame? ValidTimeFrame = TimeFrame.Forever;
+
+        /// <summary>
+        ///     What should happen if a given cassette entry is expired.
+        /// </summary>
+        public ExpirationActions WhenExpired = ExpirationActions.Warn;
 
         /// <summary>
         ///     Retrieve the manual delay as a TimeSpan.