+++ EasyVCR/Handlers/VCRHandler.cs
@@ -20,9 +20,12 @@ namespace EasyVCR.Handlers
         private readonly Censors _censors;
         private readonly TimeSpan? _delay;
         private readonly IInteractionConverter _interactionConverter;
+        private readonly ConsoleFallbackLogger _logger;
         private readonly MatchRules _matchRules;
         private readonly Mode _mode;
         private readonly bool _useOriginalDelay;
+        private readonly TimeFrame _validTimeFrame;
+        private readonly ExpirationActions _whenExpired;
 
         /// <summary>
         ///     Initializes a new instance of the <see cref="VCRHandler" /> class.
@@ -36,12 +39,17 @@ namespace EasyVCR.Handlers
             InnerHandler = innerHandler;
             _cassette = cassette;
             _mode = mode;
-
             _censors = advancedSettings?.Censors ?? new Censors();
             _interactionConverter = advancedSettings?.InteractionConverter ?? new DefaultInteractionConverter();
             _matchRules = advancedSettings?.MatchRules ?? new MatchRules();
             _useOriginalDelay = advancedSettings?.SimulateDelay ?? false;
             _delay = advancedSettings?.ManualDelayTimeSpan ?? TimeSpan.Zero;
+            _validTimeFrame = advancedSettings?.ValidTimeFrame ?? TimeFrame.Forever;
+            _whenExpired = advancedSettings?.WhenExpired ?? ExpirationActions.Warn;
+            _logger = new ConsoleFallbackLogger(advancedSettings?.Logger, "EasyVCR");
+
+            // Will throw an exception if an invalid settings combination is provided.
+            ExpirationActionExtensions.CheckCompatibleSettings(_whenExpired, _mode);
         }
 
         /// <summary>
@@ -67,18 +75,69 @@ namespace EasyVCR.Handlers
                     // try to get recorded request
                     var replayInteraction = await FindMatchingInteraction(request);
                     if (replayInteraction == null) throw new VCRException($"No interaction found for request {request.Method} {request.RequestUri}");
+                    if (_validTimeFrame.HasLapsed(replayInteraction.RecordedAt))
+                    {
+                        // matching interaction is expired
+                        switch (_whenExpired)
+                        {
+                            case ExpirationActions.Warn:
+                                // just throw a warning
+                                // will still simulate delay below
+                                _logger.Warning("Matching interaction is expired.");
+                                break;
+                            case ExpirationActions.ThrowException:
+                                // throw an exception and exit this function
+                                throw new VCRException($"Matching interaction is expired.");
+                            case ExpirationActions.RecordAgain:
+                                // we should never get here, the settings check should catch this during construction
+                                throw new VCRException("Cannot re-record an expired interaction in Replay mode.");
+                            default:
+                                // we should never get here
+                                throw new ArgumentOutOfRangeException();
+                        }
+                    }
+
                     // simulate delay if configured
                     await SimulateDelay(replayInteraction, cancellationToken);
-                    // found a matching interaction, replay response
+                    // return matching interaction's response
                     return replayInteraction.Response.ToHttpResponseMessage(request);
                 case Mode.Auto:
                     // try to get recorded request
                     var autoInteraction = await FindMatchingInteraction(request);
                     if (autoInteraction != null)
                     {
+                        // found a matching interaction
+                        if (_validTimeFrame.HasLapsed(autoInteraction.RecordedAt))
+                        {
+                            // matching interaction is expired
+                            switch (_whenExpired)
+                            {
+                                case ExpirationActions.Warn:
+                                    // just throw a warning
+                                    // will still simulate delay below
+                                    _logger.Warning("Matching interaction is expired.");
+                                    break;
+                                case ExpirationActions.ThrowException:
+                                    // throw an exception and exit this function
+                                    throw new VCRException($"Matching interaction is expired.");
+                                case ExpirationActions.RecordAgain:
+                                    //  re-record over expired interaction
+                                    // this will not execute the simulated delay, but since it's making a live request, a real delay will happen.
+                                    stopwatch.Start();
+                                    var newResponse = await base.SendAsync(request, cancellationToken);
+                                    stopwatch.Stop();
+                                    await RecordRequestAndResponse(request, newResponse, stopwatch.Elapsed);
+                                    // return the new response immediately
+                                    return newResponse;
+                                default:
+                                    // we should never get here
+                                    throw new ArgumentOutOfRangeException();
+                            }
+                        }
+
                         // simulate delay if configured
                         await SimulateDelay(autoInteraction, cancellationToken);
-                        // found a matching interaction, replay response
+                        // return matching interaction's response
                         return autoInteraction.Response.ToHttpResponseMessage(request);
                     }
 
@@ -87,6 +146,7 @@ namespace EasyVCR.Handlers
                     var autoResponse = await base.SendAsync(request, cancellationToken);
                     stopwatch.Stop();
                     await RecordRequestAndResponse(request, autoResponse, stopwatch.Elapsed);
+                    // return new interaction's response
                     return autoResponse;
                 case Mode.Bypass:
                 default: