+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## Next Release
+- New feature: Set expiration time for interactions (how long since it was recorded should an interaction be considered valid)
+  - Can determine what to do if a matching interaction is considered invalid:
+    - Warn the user, but proceed with the interaction
+    - Throw an exception
+    - Automatically re-record (cannot be used in `Replay` mode)
+- New feature: Pass in a custom ILogger instance to EasyVCR to funnel log messages into your own logging setup (fallback: logs to console)
+
 ## v0.4.0 (2022-06-13)
 
 - Improvements to censoring